rgw: add rgw_auto_remove_old_shard options
Set to false by default, it prevents RGW from automatically removing
the old bucket shards after a reshard operation

Upstream: https://tracker.ceph.com/issues/42594
Fixes: OBJECT-3168

+++ src/rgw/rgw_reshard.cc
@@ -756,20 +756,22 @@ int RGWBucketReshard::execute(int num_shards, int max_op_entries,
   // best effort and don't report out an error; the lock isn't needed
   // at this point since all we're using a best effor to to remove old
   // shard objects
-  ret = store->svc()->bi->clean_index(dpp, bucket_info);
-  if (ret < 0) {
-    ldpp_dout(dpp, -1) << "Error: " << __func__ <<
-      " failed to clean up old shards; " <<
-      "RGWRados::clean_bucket_index returned " << ret << dendl;
-  }
+  if (store->ctx()->_conf.get_val<bool>("rgw_auto_remove_old_shards")) {
+    ret = store->svc()->bi->clean_index(dpp, bucket_info);
+    if (ret < 0) {
+      ldpp_dout(dpp, -1) << "Error: " << __func__ <<
+        " failed to clean up old shards; " <<
+        "RGWRados::clean_bucket_index returned " << ret << dendl;
+    }
 
-  ret = store->ctl()->bucket->remove_bucket_instance_info(bucket_info.bucket,
-                                                       bucket_info, null_yield, dpp);
-  if (ret < 0) {
-    ldpp_dout(dpp, -1) << "Error: " << __func__ <<
-      " failed to clean old bucket info object \"" <<
-      bucket_info.bucket.get_key() <<
-      "\"created after successful resharding with error " << ret << dendl;
+    ret = store->ctl()->bucket->remove_bucket_instance_info(bucket_info.bucket,
+                                                         bucket_info, null_yield, dpp);
+    if (ret < 0) {
+      ldpp_dout(dpp, -1) << "Error: " << __func__ <<
+        " failed to clean old bucket info object \"" <<
+        bucket_info.bucket.get_key() <<
+        "\"created after successful resharding with error " << ret << dendl;
+    }
   }
 
   ldpp_dout(dpp, 1) << __func__ <<
